pyproject.toml
[tool.poetry]
name = "nyrkes"
version = "0.1.0"
description = "It is a secret!"
authors = ["Tapio Jokinen <tapio.jokinen@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
python-environ = "^0.4.54"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.13.0"
redis = "^4.4.0"
pillow = "^9.3.0"
celery = "^5.2.7"
django-celery-results = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.10.0"
isort = "^5.10.1"
flake8 = "^5.0.4"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.6.3"
pylint = "^2.15.5"
coverage = "^6.5.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[black]
line-length = 119

[tool.isort]
profile = "black"

[tool.pylint.messages_control]
disable = [
    "missing-docstring",            # We don't need this
    "line-too-long",                # Black handles this.
    "duplicate-code",               # This is judged manually.
    "abstract-method",              # This is judged manually.
    "too-few-public-methods",       # Doesn't work well with models.
    "invalid-name",                 # For e.g. OrganizationMember = apps.get_model("nyrkes", "OrganizationMember").
    "wildcard-import",              # Used in settings files.
    "unused-wildcard-import"        # Used in settings files.
]
